refactor(subnets): refactored VLANs structure
