+++ src/app/subnets/views/VLANs/components/VLANDeleteForm/VLANDeleteForm.tsx
