+++ src/app/subnets/views/VLANs/components/VLANDeleteForm/index.ts
