+++ src/app/subnets/views/VLANs/views/VLANDetails/VLANDetails.test.tsx
