+++ src/app/subnets/views/VLANs/views/VLANDetails/index.ts
