+++ src/app/subnets/views/VLANs/views/constants.ts
